Adding more juice to the ProcStatus collector
The fullerite collector ProcStatus now adds by default the CPUTime
metric. It also now has support for custom dimensions on your processes.
Imagine you want to add a dimention that is the part of the command
running the process. You need to get the data from /proc/PID/cmdline
and parse that output to extract the data you need from it.

This can now be done through your fullerite config file as per the needs
mentioned in issue #55, however in a different format.:-
{
....
"collectors": {
  "ProcStatus": {
    generatedDimensions:{
      "stream": ["cmdline", "REGEXP_GOES_HERE"],
      "somethingelse": ["....."],
    }
  },
},

cmdline is one example of a proc attribute that we support, I also
support running data against `filedescriptors`, we can extrend this
further to support more and more as we wish, we just need to handle it
in the procstatus collector.