Adidng more metrics to Scribe collector
A misisng metric for the Scribe collector was the scribe buffer size,
this can be done now by defining a `path` entry and an optional exclude_pattern
in the config file poiting to the scribe_buffer path. This defaults to None and
will hence not bne reported unless that path is defined.