Add dimensions reporting from diamond collectors
This controversial commit adds the ability to define dimensions in the
diamond collectors. The reason this is being done the way it is `as a
collector attribute` is to make the collectors still behve in their plug
and play nature. If you plug in a collector and you collectorpy does not
support dimensions, well who cares it is just an extra instance variable
that will not affect anything. If you want to catch the result of that
value to send it through to any backend well, you have it in memory. So
it's a win-win!