This change can be introduced after pull/88
Since we can add dsimensions to metrics as of pull/88 we can add this
collector easily and make it work as intended. The value of 1 for this
metric is intended as yo make it possible to sum all the timeseries
produced by the metric name and be able to get a count for each
os_distro in a distributed system.