This commit is based on pull/87
As per the standrads we hold ourselves to of testing our software, now
we can run tests on the core diamond code for colelctors and metrics.
Running the tests for these modules is done with the command:-

`python test.py -d`

This runs the tests for the diamond core *NOT* the collectors. The
colelctors are handles in pull/87.