Monitor collectors taking too long to run
We need to be able to monitor collectors taking a longer time to run
than their set run interval.
This is important so we can set a detector for this in SignalFx for
example if a collector takes too long for X colletion rounds and get
paged about this so we can jump on the host and see where this is happening and why.

Since we are adding a lot! of new collectors this is crucial so we
can easily tweak the collection interval on collector level.