Add custom dimensions to smem metrics by parsing /proc/PID/cmdline
I need to add 2 custom dimensions for Apache and uWSGI metrics: the
worker id and the process type (master, worker, http, mule).
These info can be obtained from /proc/PID/cmdline and are the same you
can see when running ps aux.

This PR allows users to extract those values and add them as dimensions.

I tested it on my machine using: "dimensionsFromCmdline":{"worker_id":
".* worker ([0-9]+)$","process_type":".* uWSGI ([^ ]+).*"}}.

This change will NOT change the number of datapoints emitted, but it
will create more timeseries. However the worker_id is always a number
between 0 and N and is reused when a worker gets restarted, so this won't
generate an infinite number of timeseries.