Add /proc/<pid>/fd collector
This is a diamond collector that counts the number of file descriptors a
certain pid is using. It finds the pid for each service specified in the
config file by looking at a pid file also specified in the config.

It could be extended to collect other data from /proc/<pid>/*