Get vagrant environment working again; add mutliple-collector example conf
The Vagrantfile was left behind after some changes to config/config.go. This
gets it working again, by:

* Updating to go 1.7.3
* Including pip in the dev environment, so diamond tests work
* Rejigged the /etc/fullerite.conf to reflect config.go changes
* added /etc/fullerite/conf.d as per ^^

So this now works:

    vagrant up
    vagrant ssh
    cd /vagrant
    make
    bin/fullerite

I was also investigating running multiple instances of the same type of
collector (with different config), which turns out is already implemented. So
i've also made the vagrant environment highlight this (via two instances of
TestCollector)