New YamlMetrics Collector
This collector gets YAML (or JSON) from a source and returns metrics for each
k/v found at the top level that can be converted to a float (including
booleans)

== Add yamlMetricsKeyWhitelist option to Yaml collector

When reading arbitrary YAML files - like the facter/mco /etc/mcollective/facts.yaml file -
it's necessary to whitelist certain keys, to prevent processing what could be
quite a lot of data and creating unnecessary/confusing signalfx DPs.

== Add boolean support to YamlMetrics

stringy or proper boolean values in Yaml are now converted to 1 (true) and
0 (false).

== Fix up glide.* to import yaml libs; fix annoying case issue with logrus

Having two different case-versions of sirupsen/logrus (vs Sirupsen/logrus)
was really freaking out glide. I could not get it to bring in the new yaml libs
or generally play nice.

== Add JSON tests, as we get it for free with YAML!

== Add 'exec' and 'shell' sources for YAML.

Prefix a yamlSource with 'exec:' to cause the command to be run directly.

Prefix a yamlSource with 'shell:' to run it under 'sh -c', so that pipelines
and args are functional.

Commands must output valid YAML (or JSON).