Rework around also supporting 'fullerite' format, with dimensions
THe 'simple' format (so we can read arbitrary YAML) can't really support
metrics with dimensions. There's reasonable prior art in the adhoc collector
to provide a specfic format that does.

This has been enabled via 'format: fullerite' in the YAML, which then reads
the "metrics" key as an array of metrics, including metrictype and dimensions.
It also means the same metric name can be specified with different dimensions.

In the process, rework and DRY the tests.