Make yamlFormat a config param, not part of YAML data. switch default to 'fullerite'
Realised that specifying the 'format: fullerite' in the YAML provided to the
collector was a daft idea. It's better to specify it in the collector
configuration.

So now we have 'yamlFormat' as a configuration option. Woot!

This allows us to set 'fullerite' as the default format, which is safer - folk
are far less likely to spit out random DPs with a structured format, and it's
the one we want folk to use.

Now that we don't have to specify the 'format' key, we're able to make this
compatible with the adhoc collector too (which just requires an array metrics).
Hopefully that aids development of new shell collectors.